fixed compiler warnings

+++ src/Pt/regexp.cpp
@@ -1293,12 +1293,12 @@ regdump( pt_regexp *r )
     s = r->program + 1;
     while (op != END) {    /* While that wasn't END last time... */
         op = OP(s);
-        printf("%2d%s", s-r->program, regprop(s));    /* Where, what. */
+        printf("%ld%s", static_cast<long>(s - r->program), regprop(s));
         next = regnext(s);
         if (next == NULL)        /* Next ptr. */
             printf("(0)");
         else
-            printf("(%d)", (s-r->program)+(next-s));
+            printf("(%ld)", static_cast<long>((s - r->program) + (next-s)));
         s += 3;
         if (op == ANYOF || op == ANYBUT || op == EXACTLY) {
             /* Literal string, where present. */
